mobile-version/requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
pydantic
pydantic-settings
pyyaml
requests
imap-tools
python-telegram-bot
python-dotenv
